+++ lib/src/dom_tools_track.dart
@@ -341,3 +341,5 @@ class TrackElementResize {
     _trackElementValueInstance.checkElements();
   }
 }
+
+